fix: PageContainer should not contain index page
PageContainer is a component that wraps around every page in the site,
but it should not be used in the index page. The index page has a
different layout from the other pages, with a fixed size that ocuppies
the user's entire window.

The PageContainer component had a fixed size of 100vh, creating a bug in
all pages besides the index. That bug made the footer being placed in
wrong position.